feat!: separate dotnet type hints and actual libraries
As automatic detection of the libraries often failed and made
testing very difficult the code was reorganised to require an
explicit import of the .NET libraries. This way importing the
PIconnect library can always succeed, regardless of the
environment. The major downside is that it requires an explicit
`PI.load_sdk()` call in every script.